tools: alphabetize ESLint rules
I have never found it useful that the ESLint rules (in the
configuration file) are organized by the categorization the rules are
given in the ESLint documentation. It has only been a source of
annoyance as one has to look up where the rule is in the docs to figure
out where it should go in the configuration file.

This change re-orders the rules so they are in alphabetical order.
Simple!

PR-URL: https://github.com/nodejs/node/pull/19100
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>
Reviewed-By: Teddy Katz <teddy.katz@gmail.com>
Reviewed-By: Roman Reiss <me@silverwind.io>
Reviewed-By: Yuta Hiroto <hello@about-hiroppy.com>
Reviewed-By: Minwoo Jung <minwoo@nodesource.com>
Reviewed-By: Shingo Inoue <leko.noor@gmail.com>
Reviewed-By: Tobias Nießen <tniessen@tnie.de>
Reviewed-By: James M Snell <jasnell@gmail.com>